move class, fix tests
