+++ pandas/tests/extension/arrow/arrays.py
@@ -24,7 +24,7 @@ from pandas.api.extensions import (
 )
 from pandas.api.types import is_scalar
 from pandas.core.arraylike import OpsMixin
-from pandas.core.arrays._arrow_utils import ArrowExtensionArray as _ArrowExtensionArray
+from pandas.core.arrays._mixins import ArrowExtensionArray as _ArrowExtensionArray
 from pandas.core.construction import extract_array
